# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for music management.",
        "title": "Music API",
        "contact": {
            "name": "Developer",
            "email": "smagulmyrzakhmet@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3030",
    "basePath": "/api/v1",
    "paths": {
        "/songs": {
            "get": {
                "description": "Получить песни из библиотеки. Можно фильтровать по всем полям кроме id\nструктуры. Есть пагинация, если не устанавливать страницу\nи лимит на количество песен будут использованы стандартные значения.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Получить песню из библиотеки",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Лимит на количество элементов на одной странице",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Название песни, фильтрация без учета регистра",
                        "name": "song",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Имя исполнителя или название группы, фильтрация без учета регистра",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Ссылка, фильтрация с учетом регистра",
                        "name": "link",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата выпуска с этой даты включительно",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата выпуска до этой даты включительно",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список песен с пагинацией и фильтрами",
                        "schema": {
                            "$ref": "#/definitions/musicmax.SongsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/musicmax.Description"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/musicmax.Description"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавить песню в библиотке",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Добавить песню в библиотке",
                "parameters": [
                    {
                        "description": "Тело запроса",
                        "name": "song",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/musicmax.SongRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/musicmax.Description"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/musicmax.Description"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/musicmax.Description"
                        }
                    }
                }
            }
        },
        "/songs/{id}": {
            "delete": {
                "description": "Удалить песню из библиотеки по id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Удалить песню из библиотеки",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/musicmax.Description"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/musicmax.Description"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/musicmax.Description"
                        }
                    }
                }
            },
            "patch": {
                "description": "Изменение данных песни",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Изменения данных песни",
                "parameters": [
                    {
                        "description": "Тело запроса",
                        "name": "songInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/musicmax.SongPatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/musicmax.Description"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/musicmax.Description"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/musicmax.Description"
                        }
                    }
                }
            }
        },
        "/songs/{id}/lyrics": {
            "get": {
                "description": "Получить текст песни по id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Получить текст песни",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Song ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Текст песни с пагинацией по куплетам",
                        "schema": {
                            "$ref": "#/definitions/musicmax.LyricsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/musicmax.Description"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/musicmax.Description"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "musicmax.Description": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        },
        "musicmax.LyricsResponse": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "song": {
                    "type": "string"
                },
                "text": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "musicmax.SongPatchRequest": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "releaseDate": {
                    "type": "string"
                },
                "song": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "musicmax.SongRequest": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "song": {
                    "type": "string"
                }
            }
        },
        "musicmax.SongResponse": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "releaseDate": {
                    "type": "string"
                },
                "song": {
                    "type": "string"
                }
            }
        },
        "musicmax.SongsResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "songs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/musicmax.SongResponse"
                    }
                }
            }
        }
    }
}